curl http stack: split up request class
